requirements.txt
discord.py==2.3.2
libscrc==1.8.1
python-dotenv==1.0.0
qrcode==7.4.2
